feat: move emails inbound method to emails.receiving.<method> (#666)
