+++ src/emails/receiving/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './get-inbound-email.interface';
