--- src/inbound/interfaces/index.ts
@@ -1,5 +1,0 @@
-export type {
-  GetInboundEmailResponse,
-  GetInboundEmailResponseSuccess,
-} from './get-inbound-email.interface';
-export type { InboundEmail, InboundEmailAttachment } from './inbound-email';
