+++ src/index.ts
@@ -7,6 +7,6 @@ export * from './common/interfaces';
 export * from './contacts/interfaces';
 export * from './domains/interfaces';
 export * from './emails/interfaces';
-export * from './inbound/interfaces';
+export * from './emails/receiving/interfaces';
 export { ErrorResponse } from './interfaces';
 export { Resend } from './resend';
